+++ view/tws_administration_view/lib/core/router/twsa_route_tree.dart
@@ -29,7 +29,7 @@ typedef Routes = TWSARoutes;
 class TWSARouteTree extends CSMRouterTreeBase {
   TWSARouteTree()
       : super(
-          devRoute: Routes.driversCreateWhisper,
+          devRoute: Routes.yardlogPage,
           redirect: (_, __) {
             return null;
           },